MDEV-18021: Galera test galera_sst_mariabackup_table_options fails if AES_CTR is not available
Problem is that if you use bundled yassl AES_CTR is not supported. There is a way to detect that but as we really want to keep this test enabled did not add
skip for missing support. Changed method to AES_CBC as there is no need to
use AES_CTR.