MDEV-17441: Pointer indirection for buf_block_t::debug_latch
This is a follow-up to commit 9581c4a8f5d6c87a6d34c6d3b826c32bf7f15143
which added a memcpy() call to rtr_copy_buf(), to copy
rw_lock_t debug_latch. This could emit a warning.

A cleaner approach is to make buf_block_t::debug_latch a pointer,
so that we can avoid copying it. An even cleaner approach would be to
redesign the InnoDB SPATIAL INDEX code so that the function
rtr_copy_buf() is not needed at all.