Remove fil_node_t::sync_event
I have a discussion with kernel fellows and they said it's safe to call fsync()
simultaneously at least on VFS and ext4. So initially I wanted to disable check
for recent Linux but than I realized code is buggy.

Consider a case when one thread is inside fsync() and two others are waining
inside os_event. First thread after fsync() calls os_event_set() which is a
broadcast! So two waiting threads will awake and may call fsync() at the same
time.

One fix is to add a notify_one() functionality to os_event but I decided to
remove incorrect check completely. Note, it works for one wainting thread but
not for more than one.

IMO it's ok to avoid existing bugs but there is not too much sense in avoiding
possible(!) bugs as this code does.

Resulting code could be a bit faster but I haven't benchmark it.