Fixes to galera_defaults, galera#505. Disabled wsrep.variables
The change introduced to galera_defaults in commit

  6b81883170e8d54d8c2efdb1a6b5b37df43e4b6d

has been lost in the merge process. Checked out the test from 10.4
head and recorded to match Galera 4 defaults.

Recorded galera#505 which is now passing. Disabled wsrep.variables
which causes a crash when global wsrep_on is manipulated during the
test in unsupported manner.