Fix crash in joined -> donor transition
It is possible that the server in joined state becomes a donor without
being notified about synced state in between. Earlier this was not
allowed by wsrep-lib which caused a crash in such a case.

This patch updates the wsrep-lib position to the point where
joiner -> donor state transition is allowed and the state change
sanity check is relaxed so that the server built in release mode
does not crash. In debug build the assertion will fail to catch
invalid api use early on.