MDEV-18515: The wsrep_cluster_weight status variable is not supported
Current version of Galera does not support the wsrep_cluster_weight
status variable, which causes the galera_pc_weight test (from the
galera_3nodes suite) to fail.

This patch adds the missing variable to the wsrep status variables.

This patch also contains improvements for other tests, which taken
from the 10.2 branch. This changes synchronize the 10.3 branch with
the latest version of the previous MDEV-18426 fixes, which are also
related to the galera_3nodes suite (which originally were made for
the 10.2 branch, and has not yet been transferred to 10.3).

https://jira.mariadb.org/browse/MDEV-18515