MDEV-15528: Introduce MLOG_INIT_FREE_PAGE
This is only a placeholder that allows an implementation later
during the development of MariaDB, so that downgrade to an earlier
version (with this code) will be possible.
We want to be able to zero out freed pages to reduce write amplification,
and to scrub old data. Zeroing out the pages is optional, not mandatory
for correctness. After all, the MLOG_INIT_FREE_PAGE record can only be
emitted for pages that are marked free in the allocation bitmap page.