MTR tests for clearing orphaned SR transactions in Galera
Added a couple of MTR tests which check that streaming transactions
get rolled back when their owners leave the primary component.
Specifically, the tests cover a case in which the cluster moves
between two primary views which have the same membership. This case is
currently not handled properly, the new version of the `wsrep-lib`
submodule addresses that.